feat: create movie routes and validation

+++ controllers/movies.js
@@ -2,10 +2,11 @@ const Movie = require('../models/movie');
 const BadRequestError = require('../errors/BadRequestError');
 const DefaultError = require('../errors/DefaultError');
 const NotFoundError = require('../errors/NotFoundError');
-
+const ForbiddenError = require('../errors/ForbiddenError');
 const {
   movieIdMissing,
   badValue,
+  ownerRigthsErr,
 } = require('../utils/errorMessages');
 
 module.exports.getMovies = (req, res, next) => {
@@ -31,6 +32,7 @@ module.exports.createMovie = (req, res, next) => {
     nameEN,
     thumbnail,
     movieId,
+
   } = req.body;
 
   Movie.create({
@@ -45,6 +47,7 @@ module.exports.createMovie = (req, res, next) => {
     nameEN,
     thumbnail,
     movieId,
+    owner: req.user._id,
   })
     .then((movie) => res.send({ data: movie }))
     .catch((err) => {
@@ -52,19 +55,51 @@ module.exports.createMovie = (req, res, next) => {
     });
 };
 
-module.exports.deleteCard = (req, res, next) => {
-  Movie.findByIdAndRemove(req.params.id)
+module.exports.deleteMovie = (req, res, next) => {
+  Movie.findById(req.params.id)
     .orFail(new Error('UnknownId'))
     .then((mov) => {
-      res.send({ message: `Фильм "${mov.name}" удален!` });
+      if (String(mov.owner) !== req.user._id) {
+        next(new ForbiddenError(ownerRigthsErr));
+        return;
+      }
+      Movie.findByIdAndRemove(req.params.id)
+        .orFail(new Error('UnknownId'))
+        .then((movie) => {
+          res.send({ message: `Фильм "${movie.nameRU}" удален!` });
+        })
+        .catch((err) => {
+          if (err.message === 'UnknownId') {
+            next(new NotFoundError(movieIdMissing));
+          } else if (err.name === 'CastError') {
+            next(new BadRequestError(badValue));
+          } else {
+            next(new DefaultError(err.message));
+          }
+        });
     })
     .catch((err) => {
       if (err.message === 'UnknownId') {
         next(new NotFoundError(movieIdMissing));
-      } else if (err.name === 'CastError') {
-        next(new BadRequestError(badValue));
       } else {
         next(new DefaultError(err.message));
       }
     });
 };
+
+// module.exports.deleteMovie = (req, res, next) => {
+//   Movie.findByIdAndRemove(req.params.movieId)
+//     .orFail(new Error('UnknownId'))
+//     .then((mov) => {
+//       res.send({ message: `Фильм "${mov.nameRU}" удален!` });
+//     })
+//     .catch((err) => {
+//       if (err.message === 'UnknownId') {
+//         next(new NotFoundError(movieIdMissing));
+//       } else if (err.name === 'CastError') {
+//         next(new BadRequestError(badValue));
+//       } else {
+//         next(new DefaultError(err.message));
+//       }
+//     });
+// };
